Move incentive card code into its own file
state-incentive-details.tsx was getting very long (it still is kinda
longer than I'd like) so factor out a self-contained subtree of it.
This is just cut-and-paste, no substantive changes.
